Added support for non-seekable streams
In some cases, where contents taken not from file but through system process pipelining STDIN is not seekable. In this case we may see a PHP warning: "rewind(): stream does not support seeking". Now we checking is stream seekable before do rewind. And we don't do rewind if stream not read yet. This allow us to read not seekable streams only once. Attempt to do rewind on not seekable stream will raise an exception with message "Cannot rewind not seekable stream".